+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Presenter/MainFormPresenter.swift
@@ -410,10 +410,13 @@ extension MainFormPresenter {
             return
         }
 
-        let isCvcValid = dataState.hasCards ? savedCardPresenter.isValid : true
-        let isEmailValid = getReceiptSwitchPresenter.isOn ? emailPresenter.isEmailValid : true
-
-        payButtonPresenter.set(enabled: isCvcValid && isEmailValid)
+        if !dataState.hasCards {
+            payButtonPresenter.set(enabled: true)
+        } else {
+            let isCvcValid = savedCardPresenter.isValid
+            let isEmailValid = getReceiptSwitchPresenter.isOn ? emailPresenter.isEmailValid : true
+            payButtonPresenter.set(enabled: isCvcValid && isEmailValid)
+        }
     }
 
     private func startPaymentWithSavedCard() {